Extract shared logic into `core`
It might be a bit easier to reason about where the logic lives
if its in a common module `core`, rather than scattered thought the
project.
